requirements.txt
beautifulsoup4==4.5.3
Django==1.10.5
django-filter==1.0.1
django-modelcluster==2.0
django-money==0.9.1
django-taggit==0.18.3
django-treebeard==4.1.0
djangorestframework==3.5.3
html5lib==0.999999999
olefile==0.44
Pillow==4.0.0
psycopg2==2.6.2
py-moneyed==0.6.0
pytz==2016.10
requests==2.12.4
six==1.10.0
sqlparse==0.2.2
Unidecode==0.4.19
wagtail==1.8
wagtailfontawesome==1.0.6
webencodings==0.5
Willow==0.4
